autotest/selenium/locations_tests/requirements.txt
pytest
selenium
mysql-connector